refactor: components/signup 폴더명 components/auth로 변경
